chore(tile-debug): move tileValidationData to separate file
